fix(symcache): Plug soundness hole (#958)
Arch is a type for which Pod is not sound to implement because not every pattern of 32 bits is valid for it.

+++ symbolic-symcache/src/writer.rs
@@ -489,7 +489,7 @@ impl<'a> SymCacheConverter<'a> {
             version: crate::SYMCACHE_VERSION,
 
             debug_id: self.debug_id,
-            arch: self.arch,
+            arch: self.arch as u32,
 
             num_files,
             num_functions,
